Optimize vp9_optimize_b slightly
Currently, at the start of this function there's a loop that performs an
indirect opaque array access through an index from an indirect opaque
array access through an index of an indirect opaque array access.
As a result 20% of this function's runtime, is spent in this one loop.

The `vp9_get_energy_class` replaces all the array reads with a
pre-computed look-up table, packed into a single value as to avoid
more array reads.

Moreover, instead of using the scan table, we can unconditionally fill
the entire token cache using `default_eob` to get the bounds of the
qcoeff array. This allows for clang to autovectorize the function.
Unfortunately gcc is somewhat resistant to my attempts at tuning
the code.

Change-Id: I5f84c560178d2190d6f2b150e212303ba5c8461c