[SSE4_1] Fix overflow in highbd temporal_filter
While porting this function to NEON, using SSE4_1 implementation as base
I noticed that both were producing files with different checksums to the
C reference implementation. After investigating further I found that this
saturating pack was the culprit. Doing the multiplication on the 32-bit
values, leads to producing the correct results with the C implementation.

Change-Id: I40c2a36551b2db363a58ea9aa19ef327f2676de3