L2E: Allow long gop sizes
We should allow a long gop size that is larger than
max_gf_interval or lag_in_frames.
But when the gop size is larger than either of the two, use_alt_ref
should be 0.
We still keep the check of use_alt_ref and return error is the
constraint is violated.

Change-Id: I5c9d7610483570742bd1367c68520b12231bb3dd