Change UpdateRateControl to return vpx_codec_err_t
Change the VP9RateControlRtcConfig constructor to initialize
ss_number_layers (to 1).

Change UpdateRateControl() to return vpx_codec_err_t so that it can
report failure (due to invalid configuration).

Also change InitRateControl() to return vpx_codec_err_t to propagate the
return value of UpdateRateControl().

Note: This is a port of the libaom CL
https://aomedia-review.googlesource.com/c/aom/+/172042.

Change-Id: I10d92f6002ff20aa3def90ac9be448790bb00ea7