Make everything into TplBlockStats
This will make the TplFrameStats just a list of TplBlockStats, instead
of a struct previously. Thus we can serialize and put them in a single
buffer, which only contains primitive types. In that way, this can be a
type of stats packet.

Also make motion vector as mv_r and mv_c.

Bug: b/273736974
Change-Id: I864e831bb10cd98ac027599eddf55728bb71f988