Do not use ml for var rd partition for INTRA ref
Because the ref frame for a certain block is set to be LAST or ARF in
ml_predict_var_rd_partitioning. When the ref frame was set to INTRA
previously in mode search, it'll get overwritten and could be an issue
when interp_filter is SWITCHABLE_FILTERS, which is legal with INTRA ref
frame.

Bug: b/290499385
Change-Id: Ia655e262f3b32445347ae0aaf1a2d868cea997f3