rt: Force nonrd_keyframe under dynamic change of deadline
For realtime mode: if the deadline mode (good/best/realtime)
is changed on the fly (via codec_encode() call), force a
key frame and set nonrd_keyframe = 1 to avoid entering the
rd pickmode.

Dynamic change of deadline, in particular for realtime mode,
involves a lot of coding feature changes, so best to force
reset with keyframe and also avoid rd pickmode on that keyframe.

b/310663186

Change-Id: Idf8fd7c9ee54b301968184be5481ee9faa06468d