Add test for issue: b/329088759
Add test for ROW_MT=0 and 1, both reproduce
the issue in the bug below. For ROW_MT=1 we
need to set 2 threads instead of 0 for the
middle config calls.

Test possible fix: change the condition to
re-alloc bitstream_worker_data to force re-alloc
if current dest_size is smaller than the current
frame size.

Bug: b/329088759
Change-Id: Icd65dbc5317120304d803f648d4bd9405710db6f