Fix and unittest added for issue: b/329088759
Add test for ROW_MT=0 and 1, both reproduce
the issue in the bug below. For ROW_MT=1 we
need to set 2 threads instead of 0 for the
middle config calls.

The failures in the test were due to using the
bitstream_worker_data when it wasn't allocated for
big enough size. This is because the existing
condition was to only re-alloc the bitstream_worker_data
when current dest_size was larger than the current frame_size.
But under resolution change where frame_size is increased,
beyond the current dest_size, we need to allow re-alloc
to the new size.

The existing condition to re-alloc when dest_size is
larger than frame_size (which is not required) is kept
for now.

Also increasee the dest_size to account for UV and
frame header offset.

Bug: b/329088759
Change-Id: Icd65dbc5317120304d803f648d4bd9405710db6f