Fix high target data rate overflow.
These change fixes issues that can occur if the user specifies a very
high target data rate or rate per frame.

Fixes some issue with overflow of int variables used to hold bitrate
values (rate per second, rate per frame etc).

This patch also imposes a new maximum for the passed in target bitrate.
This value is passed in in kbits (so multiplied by 1000
internally).

Note: This CL is a port of the following libaom CLs:
https://aomedia-review.googlesource.com/c/aom/+/190381
https://aomedia-review.googlesource.com/c/aom/+/190462

All the changes were ported to VP9. For VP8, only the new maximum for
rc_target_bitrate and the new type of cpi->bytes (equivalent to
ppi->total_bytes in libaom) were ported.

Change-Id: I438dd46efd5a134389b893ffae1f8a2381207906