Fix further overflow issue in VBR.
This patch fixes some additional cases where under extreme conditions
some of the VBR adjustment variables can wrap.

As this happens on a per frame level the extra saturation checks should
not be an issue for performance.

Note: This CL is a port of the libaom CL
https: //aomedia-review.googlesource.com/c/aom/+/190521.
Change-Id: I87c4ecca10f39767002f7d90d0f43b19c7150832