vp8: fix OOB access in x->MVcount
best_ref_mv is now clamped in vp8_rd_pick_inter_mode() and
vp8_pick_inter_mode() to avoid issues in rd_update_mvcount() and
update_mvcount(), and to avoid propagating the large motion vectors
which can cause rollovers for short values elsewhere. The difference of
valid motion vectors may still go beyond the range of the MVcount array,
however, so additional checks are added to rd_update_mvcount() and
update_mvcount().

Note the test source and settings (speed 1 and GOOD quality mode) come
from the issue report; additional coverage is added for realtime. The
realtime path does not trigger the error without the fix, but as it's
similar to the rd path, the same clamp is done to be safe.

Fixes:
vp8/encoder/rdopt.c:1579:5: runtime error: index 17467 out of bounds for
  type 'unsigned int[2047]'

Bug: oss-fuzz:69906
Change-Id: Ia8bd087cfe4475ab09ba711ed806fbcbaa72e552