rtc-vp9: Re-enable svc_use_low_part for some conditons
Avoid entering the function scale_partitioning_svc()
near frame boundary.

This is a possible fix for issue below. By re-enabling
the feature under these conditions we get most of the
~3% speedup back (for (3SL,3TL) encoding).

Added unittest that triggers (without this fix) an assert
in vp9_bitstream.c:423: write_partition:
Assertion `p == PARTITION_SPLIT' failed.

The added test uses a clip of resolution 1284x770, otherwise
the settings are the same as an existing 720p test.

Bug: 366146260
Change-Id: I199e225dd76d3fbe93d814a436ef2cdcf12bb36d