rtc-vp9: Re-enable svc_use_low_part for some conditons
Avoid entering the function scale_partitioning_svc()
near frame boundary and for small blocks (< 16x16).

This is a possible fix for issue below. By re-enabling
the feature under these conditions we get most of the
speedup back. It's likely the crash was caused along
the boundary or for small blocks, though we cannot repro.

Bug: 366146260
Change-Id: I199e225dd76d3fbe93d814a436ef2cdcf12bb36d