rtc-vp9: Re-enable svc_use_low_part for some conditons
Avoid entering the function scale_partitioning_svc()
near frame boundary. And incorporate the (bit-exact)
change (to remove unneeded variables) from:
https://chromium-review.googlesource.com/c/webm/libvpx/+/5873446

Also fix an integer overflow caused by the newly added test.

This is a possible fix for issue below. By re-enabling
the feature under these conditions we get most of the
~3% speedup back (for (3SL,3TL) encoding).

This CL undoes some of the changes in:
https://chromium-review.googlesource.com/c/webm/libvpx/+/768508

Added unittest that triggers (without this fix of avoiding
the boundary) an assert in vp9_bitstream.c:423: write_partition:
Assertion `p == PARTITION_SPLIT' failed.

The added test uses a clip of resolution 1284x770, otherwise
the settings are the same as an existing 720p test.

Bug: 366146260
Change-Id: I199e225dd76d3fbe93d814a436ef2cdcf12bb36d