rtc: Fix issues with use_base_mv in pickmode
For the use_base_mv svc feature in vp9_pickmode:
check for integer overflow and out of bounds
for array prev_frame->mvs[]. For the mvs array
keep track of prev_frame  mi_col/rows, to be
used to access and check the low-resol/prev mv.

These issues are picked up by the added svc
test for 1284x770 resolution.

Change-Id: Idab6edf9f7f3b87a86c71c33913bd9e80aec6307