rtc: Fixes to scale_partitioning_svc()
The feature scale_partitioning_svc() is controlled
by the speed feature sf->svc_use_lowres_part, which is
currently disabled.

This CL keeps the feature disabled, but fixes issues in
scale_partitioning_svc(), when enabled, picked up by the
new added tests in:
https://chromium-review.googlesource.com/c/webm/libvpx/+/5919306

The fixes are:
1) avoid entering the function scale_partitioning_svc()
near frame boundary
2) return 1 (do variance partitioning) instead of return 0
(do scale partitioning) when the lower-resolution frame is
outside its boundary.
3) incorporate the (bit-exact) change (to remove unneeded variables):
https://chromium-review.googlesource.com/c/webm/libvpx/+/5873446

This CL undoes some of the changes in:
https://chromium-review.googlesource.com/c/webm/libvpx/+/768508

Bug: 366146260
Change-Id: I199e225dd76d3fbe93d814a436ef2cdcf12bb36d